refactor: update new imports for IconBarbell component
