+++ src/modules/exercise/presentation/ui/components/ModalExerciseForm.tsx
@@ -1,18 +1,18 @@
 'use client';
 
-import { createExerciseAction } from '../../exercise.actions';
-import { ExerciseDescriptionField } from './fields/ExerciseDescriptionField';
-import { ExerciseFormSchema } from '../schemas/exercise.schema';
-import { ExerciseMusclesField } from './fields/ExerciseMusclesField';
-import { ExerciseNameField } from './fields/ExerciseNameField';
-import { ExerciseRepsField } from './fields/ExerciseRepsField';
-import { ExerciseSetsField } from './fields/ExerciseSetsField';
-import { ExerciseWeightField } from './fields/ExerciseWeightField';
+import { useRouter } from 'next/navigation';
+import { createExerciseAction } from '@/modules/exercise/presentation/exercise.actions';
+import { ExerciseDescriptionField } from '@/modules/exercise/presentation/ui/components/fields/ExerciseDescriptionField';
+import { ExerciseFormSchema } from '@/modules/exercise/presentation/ui/schemas/exercise.schema';
+import { ExerciseMusclesField } from '@/modules/exercise/presentation/ui/components/fields/ExerciseMusclesField';
+import { ExerciseNameField } from '@/modules/exercise/presentation/ui/components/fields/ExerciseNameField';
+import { ExerciseRepsField } from '@/modules/exercise/presentation/ui/components/fields/ExerciseRepsField';
+import { ExerciseSetsField } from '@/modules/exercise/presentation/ui/components/fields/ExerciseSetsField';
+import { ExerciseWeightField } from '@/modules/exercise/presentation/ui/components/fields/ExerciseWeightField';
 import { FormFieldSection } from '@/presentation/modules/form/components/FormFieldSection';
-import { IconBarbell } from '@/presentation/globals/components/icons/Icons';
+import { IconBarbell } from '@/presentation/globals/components/icons/outline/IconBarbell';
 import { ModalForm } from '@/presentation/modules/form/components/modal-form/ModalForm';
 import { ModalFormButtons } from '@/presentation/modules/form/components/modal-form/ModalFormButtons';
-import { useRouter } from 'next/navigation';
 import type { ExerciseDTO } from '@/modules/exercise/application/dtos/exercise.dto';
 import type { SelectOption } from '@/presentation/modules/form/form.types';
