+++ src/presentation/globals/components/day-type/TypeDayIcon.tsx
@@ -1,4 +1,5 @@
-import { IconBarbell, IconBarbellOff } from '../icons/Icons';
+import { IconBarbell } from '@/presentation/globals/components/icons/outline/IconBarbell';
+import { IconBarbellOff } from '@/presentation/globals/components/icons/outline/IconBarbellOff';
 
 export function TrainingIcon({
   className = '',
