+++ src/presentation/globals/components/icons/Icons.tsx
@@ -1,20 +1,6 @@
 import type { IconProps } from '@/presentation/globals/presentation.types';
 import { IconWrapper } from '@/presentation/globals/components/icons/IconWrapper';
 
-export function IconBarbell(props: IconProps) {
-  return (
-    <IconWrapper {...props}>
-      <path d="M2 12h1" />
-      <path d="M6 8h-2a1 1 0 0 0 -1 1v6a1 1 0 0 0 1 1h2" />
-      <path d="M6 7v10a1 1 0 0 0 1 1h1a1 1 0 0 0 1 -1v-10a1 1 0 0 0 -1 -1h-1a1 1 0 0 0 -1 1z" />
-      <path d="M9 12h6" />
-      <path d="M15 7v10a1 1 0 0 0 1 1h1a1 1 0 0 0 1 -1v-10a1 1 0 0 0 -1 -1h-1a1 1 0 0 0 -1 1z" />
-      <path d="M18 8h2a1 1 0 0 1 1 1v6a1 1 0 0 1 -1 1h-2" />
-      <path d="M22 12h-1" />
-    </IconWrapper>
-  );
-}
-
 export function IconBell(props: IconProps) {
   return (
     <IconWrapper {...props}>
