+++ src/presentation/modules/sidebar/components/SidebarMainSection.tsx
@@ -1,15 +1,13 @@
 'use client';
 
 import { SidebarSection } from '@/presentation/modules/sidebar/components/SidebarSection';
-import {
-  IconBarbell,
-  IconMan,
-  IconRouteX,
-  IconSchema,
-} from '@/presentation/globals/components/icons/Icons';
-import { NavLink } from './nav/NavLink';
-import { NavLinks } from './nav/NavLinks';
-import { NavLinkIcon } from './nav/NavLinkIcon';
+import { IconBarbell } from '@/presentation/globals/components/icons/outline/IconBarbell';
+import { IconMan } from '@/presentation/globals/components/icons/outline/IconMan';
+import { IconRouteX } from '@/presentation/globals/components/icons/outline/IconRouteX';
+import { IconSchema } from '@/presentation/globals/components/icons/outline/IconSchema';
+import { NavLink } from '@/presentation/modules/sidebar/components/nav/NavLink';
+import { NavLinks } from '@/presentation/modules/sidebar/components/nav/NavLinks';
+import { NavLinkIcon } from '@/presentation/modules/sidebar/components/nav/NavLinkIcon';
 
 export function SidebarMainSection() {
   return (
